BlockNumber use long to replace BigInteger.
Former-commit-id: ada18945794a3be2abf208eb0928ab63715a8ed3

+++ sirius-chainconnector/src/main/kotlin/org/starcoin/sirius/protocol/ethereum/InMemoryChain.kt
@@ -105,8 +105,8 @@ class InMemoryChain(autoGenblock: Boolean = true) : EthereumBaseChain() {
         }
     }
 
-    override fun getBlockNumber(): BigInteger {
-        return sb.blockchain.bestBlock.number.toBigInteger()
+    override fun getBlockNumber(): Long {
+        return sb.blockchain.bestBlock.number
     }
 
     override fun newTransaction(account: EthereumAccount,to:Address,value:BigInteger):EthereumTransaction {
