+++ sirius-core/src/main/kotlin/org/starcoin/sirius/protocol/Chain.kt
@@ -41,7 +41,7 @@ interface Chain<T : ChainTransaction, B : Block<T>, A : ChainAccount> {
 
     fun deployContract(account: A, contractFile: File, args: ContractConstructArgs): HubContract<A>
 
-    fun getBlockNumber(): BigInteger
+    fun getBlockNumber(): Long
 
     fun newTransaction(account: A, to: Address, value: BigInteger): T
 }
